Move timed_job/utils out of index
